% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_corr_samples.R
\name{boot_corr_samples}
\alias{boot_corr_samples}
\title{Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample}
\usage{
boot_corr_samples(
  data,
  simu_size = 100,
  col_index1,
  col_index2,
  by = "pearson"
)
}
\arguments{
\item{simu_size}{Must be a number. Refers to the number of Bootstrap samples to be generated}

\item{data:}{An object of type data.frame. Refers to the sample where the resampling method will be performed}

\item{col_index1:}{Must be a number. Refers to the index of the date column that the correlation is to be calculated}

\item{col_index2:}{Must be a number. Refers to the index of the date column that the correlation is to be calculated}

\item{by:}{Must be: 'pearson', 'spearman' or 'kendall'. Refers to the calculated correlation method}
}
\description{
Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample
}
