% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_index_samples.R
\name{boot_index_samples}
\alias{boot_index_samples}
\title{Given a data.frame or a matrix of N rows (observations), a matrix N X simu_size will be returned. Where each row refers to the observation index in the original object and each column is a sample}
\usage{
boot_index_samples(data, simu_size = 100)
}
\arguments{
\item{data}{An object of type data.frame or matrix. Refers to the sample where the resampling method will be performed}

\item{simu_size}{Must be a number. Refers to the number of Bootstrap samples to be generated}
}
\description{
Given a data.frame or a matrix of N rows (observations), a matrix N X simu_size will be returned. Where each row refers to the observation index in the original object and each column is a sample
}
