% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_quality.R
\name{boot_quality}
\alias{boot_quality}
\title{Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample}
\usage{
boot_quality(data, theta = mean(data))
}
\arguments{
\item{data:}{An object of type vector. Refers to the data where the range will be calculated}

\item{theta:}{An object of type numeric. Refers to the true value of the parameter to be estimated}
}
\description{
Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample
}
