% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_quantile.R
\name{boot_quantile}
\alias{boot_quantile}
\title{Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample}
\usage{
boot_quantile(data)
}
\description{
Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample
}
