% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_stat.R
\name{boot_stat}
\alias{boot_stat}
\title{Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample}
\usage{
boot_stat(data, FUN)
}
\arguments{
\item{data:}{data: An object of type data.frame. Refers to the data where the static function calculation will be performed}

\item{FUN:}{Must be a function. Refers to the statistical function to be applied on a date}
}
\description{
Given a vector of size N, the matrix N X simu_size will be returned. Where each row is a sample observation and each column is a sample
}
